Refactor resource.rs into focused modules using Extract Method pattern
The 1,415-line resource.rs file violated single responsibility principle
by containing phantom types, concrete implementations, lifecycle patterns,
monitoring, and tests in one massive file. This made the code difficult
to understand, maintain, and test.

Applied Extract Method pattern to split responsibilities:
- resource/implementations.rs: Database and locking implementations
- resource/lifecycle.rs: Resource lifecycle patterns (scoped, timed, managed)
- resource/monitor.rs: Resource monitoring and leak detection
- resource/tests.rs: All test cases moved from main file

Main resource.rs reduced from 1,415 to 269 lines while maintaining
identical public API. All integration tests continue to pass.
