+++ PLANNING.md
@@ -70,10 +70,10 @@ EventCore has successfully completed all initially planned phases (1-20), includ
 **Problem**: Single file handling all resource lifecycle patterns
 **Tasks**:
 - [x] Extract phantom type definitions to `resource/types.rs`
-- [ ] Move concrete implementations to `resource/implementations.rs`
-- [ ] Create `resource/lifecycle.rs` for acquisition/release patterns
-- [ ] Create `resource/pool.rs` for resource pooling
-- [ ] Create `resource/monitor.rs` for resource monitoring
+- [x] Move concrete implementations to `resource/implementations.rs`
+- [x] Create `resource/lifecycle.rs` for acquisition/release patterns
+- [x] Create `resource/pool.rs` for resource pooling (Note: combined with lifecycle.rs)
+- [x] Create `resource/monitor.rs` for resource monitoring
 
 #### 4. Refactor errors.rs Clone implementation (243 lines) - **MEDIUM**
 **Problem**: Massive manual Clone implementation for error types
