+++ eventcore/src/resource.rs
@@ -4,12 +4,19 @@
 //! Resources must be acquired before use and cannot be used after release.
 //! The type system prevents use-after-release and double-release errors.
 
+pub mod implementations;
+pub mod lifecycle;
+pub mod monitor;
+#[cfg(test)]
+mod tests;
 pub mod types;
 
+pub use implementations::*;
+pub use lifecycle::{ManagedResource, ResourceExt, ResourceScope, TimedResourceGuard};
+pub use monitor::{global_leak_detector, ResourceLeakDetector, ResourceLeakStats};
 pub use types::{states, IsAcquired, IsReleasable, IsRecoverable};
 
 use std::marker::PhantomData;
-use std::time::{Duration, Instant};
 
 use async_trait::async_trait;
 use thiserror::Error;
@@ -225,1114 +232,6 @@ pub trait ResourceManager<T> {
     /// Callers must transition to acquired or failed state
     fn create_initializing(inner: T) -> Resource<T, states::Initializing> {
         Resource::new(inner)
-    }
-}
-
-/// Scoped resource management with automatic cleanup
-///
-/// This ensures resources are always released when the scope ends,
-/// even if an error occurs. Supports both async and sync cleanup.
-pub struct ResourceScope<T> {
-    resource: Option<Resource<T, states::Acquired>>,
-    leaked: bool,
-}
-
-impl<T> ResourceScope<T> {
-    /// Create a new resource scope with an acquired resource
-    pub const fn new(resource: Resource<T, states::Acquired>) -> Self {
-        Self {
-            resource: Some(resource),
-            leaked: false,
-        }
-    }
-
-    /// Access the resource within the scope
-    ///
-    /// Panics if the resource has already been released
-    pub fn with_resource<F, R>(&mut self, f: F) -> R
-    where
-        F: FnOnce(&mut Resource<T, states::Acquired>) -> R,
-    {
-        let resource = self
-            .resource
-            .as_mut()
-            .expect("Resource has already been released from scope");
-        f(resource)
-    }
-
-    /// Manually release the resource from the scope
-    ///
-    /// If not called, the resource will be automatically released when dropped
-    pub fn release(mut self) -> ResourceResult<()> {
-        if let Some(resource) = self.resource.take() {
-            resource.release()?;
-        }
-        Ok(())
-    }
-
-    /// Check if the resource has been released
-    pub const fn is_released(&self) -> bool {
-        self.resource.is_none()
-    }
-
-    /// Check if the resource was leaked (dropped without explicit release)
-    pub const fn is_leaked(&self) -> bool {
-        self.leaked
-    }
-}
-
-impl<T> Drop for ResourceScope<T> {
-    fn drop(&mut self) {
-        if self.resource.is_some() {
-            self.leaked = true;
-            tracing::error!("ResourceScope dropped without explicit release - resource may leak");
-
-            // Attempt to perform emergency cleanup if the resource supports it
-            if let Some(resource) = self.resource.take() {
-                // Try to release synchronously if possible
-                // Note: This is a best-effort cleanup for resources that don't require async release
-                drop(resource);
-            }
-        }
-    }
-}
-
-/// Timeout-based resource cleanup guard
-///
-/// Automatically releases resources after a specified timeout if not explicitly released
-pub struct TimedResourceGuard<T>
-where
-    T: Send + 'static,
-{
-    resource: Option<Resource<T, states::Acquired>>,
-    timeout: Duration,
-    acquired_at: Instant,
-    cleanup_task: Option<tokio::task::JoinHandle<()>>,
-}
-
-impl<T> TimedResourceGuard<T>
-where
-    T: Send + 'static,
-{
-    /// Create a new timed resource guard
-    pub fn new(resource: Resource<T, states::Acquired>, timeout: Duration) -> Self {
-        let acquired_at = Instant::now();
-
-        Self {
-            resource: Some(resource),
-            timeout,
-            acquired_at,
-            cleanup_task: None,
-        }
-    }
-
-    /// Create a new timed resource guard with automatic cleanup task
-    pub fn new_with_auto_cleanup(resource: Resource<T, states::Acquired>, timeout: Duration) -> Self
-    where
-        T: Send + Sync + 'static,
-    {
-        let acquired_at = Instant::now();
-
-        // Note: In a real implementation, we'd need a way to cancel the cleanup task
-        // when the resource is manually released. This is a simplified version.
-
-        Self {
-            resource: Some(resource),
-            timeout,
-            acquired_at,
-            cleanup_task: None,
-        }
-    }
-
-    /// Check if the resource has timed out
-    pub fn is_timed_out(&self) -> bool {
-        self.acquired_at.elapsed() > self.timeout
-    }
-
-    /// Get time remaining before timeout
-    pub fn time_remaining(&self) -> Option<Duration> {
-        self.timeout.checked_sub(self.acquired_at.elapsed())
-    }
-
-    /// Access the resource if it hasn't timed out
-    pub fn get(&self) -> Option<&T>
-    where
-        T: Send,
-    {
-        if self.is_timed_out() {
-            None
-        } else {
-            self.resource.as_ref().map(Resource::get)
-        }
-    }
-
-    /// Release the resource manually before timeout
-    pub fn release(mut self) -> ResourceResult<()> {
-        if let Some(cleanup_task) = self.cleanup_task.take() {
-            cleanup_task.abort();
-        }
-
-        if let Some(resource) = self.resource.take() {
-            if self.is_timed_out() {
-                return Err(ResourceError::Timeout {
-                    duration: self.acquired_at.elapsed(),
-                });
-            }
-            resource.release()?;
-        }
-        Ok(())
-    }
-}
-
-impl<T> Drop for TimedResourceGuard<T>
-where
-    T: Send + 'static,
-{
-    fn drop(&mut self) {
-        if let Some(cleanup_task) = self.cleanup_task.take() {
-            cleanup_task.abort();
-        }
-
-        if self.resource.is_some() {
-            if self.is_timed_out() {
-                tracing::error!(
-                    "TimedResourceGuard dropped after timeout of {:?} - resource forcibly cleaned up",
-                    self.timeout
-                );
-            } else {
-                tracing::warn!("TimedResourceGuard dropped before timeout - resource may leak");
-            }
-        }
-    }
-}
-
-/// Resource leak detector for debugging and monitoring
-#[derive(Debug, Default)]
-pub struct ResourceLeakDetector {
-    active_resources: std::sync::Mutex<std::collections::HashMap<String, ResourceInfo>>,
-}
-
-#[derive(Debug, Clone)]
-struct ResourceInfo {
-    resource_type: String,
-    acquired_at: Instant,
-    location: Option<String>,
-}
-
-impl ResourceLeakDetector {
-    /// Create a new leak detector
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Register a resource acquisition
-    pub fn register_acquisition(
-        &self,
-        resource_id: &str,
-        resource_type: &str,
-        location: Option<String>,
-    ) {
-        if let Ok(mut resources) = self.active_resources.lock() {
-            resources.insert(
-                resource_id.to_string(),
-                ResourceInfo {
-                    resource_type: resource_type.to_string(),
-                    acquired_at: Instant::now(),
-                    location,
-                },
-            );
-        }
-    }
-
-    /// Register a resource release
-    pub fn register_release(&self, resource_id: &str) {
-        if let Ok(mut resources) = self.active_resources.lock() {
-            resources.remove(resource_id);
-        }
-    }
-
-    /// Get statistics about active resources
-    pub fn get_stats(&self) -> ResourceLeakStats {
-        self.active_resources.lock().map_or_else(
-            |_| ResourceLeakStats::default(),
-            |resources| {
-                let total_count = resources.len();
-                let mut by_type = std::collections::HashMap::new();
-                let mut oldest_age = Duration::ZERO;
-
-                for info in resources.values() {
-                    *by_type.entry(info.resource_type.clone()).or_insert(0) += 1;
-                    let age = info.acquired_at.elapsed();
-                    if age > oldest_age {
-                        oldest_age = age;
-                    }
-                }
-
-                ResourceLeakStats {
-                    total_active: total_count,
-                    by_type,
-                    oldest_resource_age: oldest_age,
-                }
-            },
-        )
-    }
-
-    /// Find potentially leaked resources (older than threshold)
-    pub fn find_potential_leaks(&self, threshold: Duration) -> Vec<String> {
-        self.active_resources.lock().map_or_else(
-            |_| Vec::new(),
-            |resources| {
-                resources
-                    .iter()
-                    .filter(|(_, info)| info.acquired_at.elapsed() > threshold)
-                    .map(|(id, _)| id.clone())
-                    .collect()
-            },
-        )
-    }
-}
-
-/// Statistics about resource usage and potential leaks
-#[derive(Debug, Default)]
-pub struct ResourceLeakStats {
-    /// Total number of active resources
-    pub total_active: usize,
-    /// Count of active resources by type
-    pub by_type: std::collections::HashMap<String, usize>,
-    /// Age of the oldest active resource
-    pub oldest_resource_age: Duration,
-}
-
-/// Global resource leak detector instance
-static GLOBAL_LEAK_DETECTOR: std::sync::OnceLock<ResourceLeakDetector> = std::sync::OnceLock::new();
-
-/// Get the global resource leak detector
-pub fn global_leak_detector() -> &'static ResourceLeakDetector {
-    GLOBAL_LEAK_DETECTOR.get_or_init(ResourceLeakDetector::new)
-}
-
-/// Automatic cleanup resource wrapper
-///
-/// This wrapper automatically registers resources for leak detection
-/// and provides cleanup on drop
-pub struct ManagedResource<T, S> {
-    inner: Option<Resource<T, S>>,
-    resource_id: String,
-    cleanup_registered: bool,
-}
-
-impl<T, S> ManagedResource<T, S> {
-    /// Create a new managed resource
-    pub fn new(resource: Resource<T, S>, resource_type: &str) -> Self {
-        let resource_id = format!(
-            "{}_{}",
-            resource_type,
-            uuid::Uuid::new_v7(uuid::Timestamp::now(uuid::NoContext))
-        );
-
-        // Register with leak detector
-        global_leak_detector().register_acquisition(
-            &resource_id,
-            resource_type,
-            Some(format!("{}:{}:{}", file!(), line!(), column!())),
-        );
-
-        Self {
-            inner: Some(resource),
-            resource_id,
-            cleanup_registered: true,
-        }
-    }
-
-    /// Get a reference to the inner resource
-    pub const fn get(&self) -> Option<&Resource<T, S>> {
-        self.inner.as_ref()
-    }
-
-    /// Take the inner resource, transferring ownership
-    pub fn take(mut self) -> Option<Resource<T, S>> {
-        if self.cleanup_registered {
-            global_leak_detector().register_release(&self.resource_id);
-            self.cleanup_registered = false;
-        }
-        self.inner.take()
-    }
-}
-
-impl<T, S> Drop for ManagedResource<T, S> {
-    fn drop(&mut self) {
-        if self.cleanup_registered {
-            global_leak_detector().register_release(&self.resource_id);
-        }
-
-        if self.inner.is_some() {
-            tracing::debug!("ManagedResource dropped with resource still present");
-        }
-    }
-}
-
-/// Extension trait for adding automatic cleanup to resources
-pub trait ResourceExt<T, S>: Sized {
-    /// Wrap in a managed resource for automatic leak detection
-    fn managed(self, resource_type: &str) -> ManagedResource<T, S>;
-
-    /// Wrap in a scoped resource for automatic cleanup
-    fn scoped(self) -> ResourceScope<T>
-    where
-        S: IsAcquired;
-
-    /// Wrap in a timed guard for timeout-based cleanup
-    fn with_timeout(self, timeout: Duration) -> TimedResourceGuard<T>
-    where
-        S: IsAcquired,
-        T: Send + 'static;
-}
-
-// Implementation only for Acquired state to avoid unsafe code and conflicts
-impl<T> ResourceExt<T, states::Acquired> for Resource<T, states::Acquired> {
-    fn managed(self, resource_type: &str) -> ManagedResource<T, states::Acquired> {
-        ManagedResource::new(self, resource_type)
-    }
-
-    fn scoped(self) -> ResourceScope<T> {
-        ResourceScope::new(self)
-    }
-
-    fn with_timeout(self, timeout: Duration) -> TimedResourceGuard<T>
-    where
-        T: Send + 'static,
-    {
-        TimedResourceGuard::new(self, timeout)
-    }
-}
-
-// Implementation for other states (can only use managed)
-impl<T, S> Resource<T, S> {
-    /// Create a managed resource for any state
-    pub fn managed(self, resource_type: &str) -> ManagedResource<T, S> {
-        ManagedResource::new(self, resource_type)
-    }
-}
-
-/// Database connection resource implementation
-#[cfg(feature = "postgres")]
-pub mod database {
-    use super::{async_trait, states, Resource, ResourceError, ResourceManager, ResourceResult};
-    use sqlx::{PgPool, Postgres, Transaction};
-    use std::sync::Arc;
-    use std::time::{Duration, Instant};
-
-    /// A database connection pool wrapped in resource management
-    pub type DatabasePool = Resource<Arc<PgPool>, states::Acquired>;
-
-    /// A database transaction wrapped in resource management
-    pub type DatabaseTransaction<'a> = Resource<Transaction<'a, Postgres>, states::Acquired>;
-
-    /// Database connection wrapped in resource management  
-    pub type DatabaseConnection = Resource<sqlx::pool::PoolConnection<Postgres>, states::Acquired>;
-
-    /// Database pool resource manager with health monitoring
-    pub struct DatabaseResourceManager {
-        pool: Arc<PgPool>,
-        last_health_check: std::sync::Mutex<Instant>,
-        health_check_interval: Duration,
-    }
-
-    impl DatabaseResourceManager {
-        /// Create a new database resource manager
-        pub fn new(pool: Arc<PgPool>) -> Self {
-            Self {
-                pool,
-                last_health_check: std::sync::Mutex::new(Instant::now()),
-                health_check_interval: Duration::from_secs(30),
-            }
-        }
-
-        /// Create a new database resource manager with custom health check interval
-        pub fn new_with_health_interval(
-            pool: Arc<PgPool>,
-            health_check_interval: Duration,
-        ) -> Self {
-            Self {
-                pool,
-                last_health_check: std::sync::Mutex::new(Instant::now()),
-                health_check_interval,
-            }
-        }
-
-        /// Check if health check is needed
-        fn needs_health_check(&self) -> bool {
-            self.last_health_check.lock().map_or(true, |last_check| {
-                last_check.elapsed() > self.health_check_interval
-            })
-        }
-
-        /// Perform health check and update timestamp
-        async fn perform_health_check(&self) -> ResourceResult<()> {
-            // Basic connectivity check
-            sqlx::query("SELECT 1")
-                .execute(self.pool.as_ref())
-                .await
-                .map_err(|e| {
-                    ResourceError::AcquisitionFailed(format!("Health check failed: {e}"))
-                })?;
-
-            // Update health check timestamp
-            if let Ok(mut last_check) = self.last_health_check.lock() {
-                *last_check = Instant::now();
-            }
-
-            Ok(())
-        }
-    }
-
-    #[async_trait]
-    impl ResourceManager<Arc<PgPool>> for DatabaseResourceManager {
-        async fn acquire() -> ResourceResult<Resource<Arc<PgPool>, states::Acquired>> {
-            // Static method - cannot access instance data
-            Err(ResourceError::AcquisitionFailed(
-                "DatabaseResourceManager::acquire requires instance method".to_string(),
-            ))
-        }
-    }
-
-    impl DatabaseResourceManager {
-        /// Acquire a database pool resource with health checking
-        pub async fn acquire_pool(&self) -> ResourceResult<DatabasePool> {
-            // Perform health check if needed
-            if self.needs_health_check() {
-                self.perform_health_check().await?;
-            }
-
-            // Verify pool is not closed
-            if self.pool.is_closed() {
-                return Err(ResourceError::AcquisitionFailed(
-                    "Connection pool is closed".to_string(),
-                ));
-            }
-
-            Ok(Resource::new(Arc::clone(&self.pool)))
-        }
-
-        /// Acquire a single database connection resource
-        pub async fn acquire_connection(&self) -> ResourceResult<DatabaseConnection> {
-            // Perform health check if needed
-            if self.needs_health_check() {
-                self.perform_health_check().await?;
-            }
-
-            // Acquire a connection from the pool
-            let connection = self.pool.acquire().await.map_err(|e| {
-                ResourceError::AcquisitionFailed(format!("Failed to acquire connection: {e}"))
-            })?;
-
-            Ok(Resource::new(connection))
-        }
-
-        /// Begin a transaction with resource management
-        pub async fn begin_transaction(&self) -> ResourceResult<DatabaseTransaction<'_>> {
-            // Perform health check if needed
-            if self.needs_health_check() {
-                self.perform_health_check().await?;
-            }
-
-            // Begin transaction
-            let transaction = self.pool.begin().await.map_err(|e| {
-                ResourceError::AcquisitionFailed(format!("Failed to begin transaction: {e}"))
-            })?;
-
-            Ok(Resource::new(transaction))
-        }
-    }
-
-    /// Extension methods for database pool resources
-    impl DatabasePool {
-        /// Execute a query using the resource
-        ///
-        /// Only available when the resource is acquired
-        pub async fn execute_query(
-            &self,
-            query: &str,
-        ) -> ResourceResult<sqlx::postgres::PgQueryResult> {
-            sqlx::query(query)
-                .execute(self.get().as_ref())
-                .await
-                .map_err(|e| ResourceError::InvalidState(format!("Query execution failed: {e}")))
-        }
-
-        /// Fetch one row from a query
-        ///
-        /// Only available when the resource is acquired
-        pub async fn fetch_one(&self, query: &str) -> ResourceResult<sqlx::postgres::PgRow> {
-            sqlx::query(query)
-                .fetch_one(self.get().as_ref())
-                .await
-                .map_err(|e| ResourceError::InvalidState(format!("Query fetch failed: {e}")))
-        }
-
-        /// Get the connection pool
-        ///
-        /// Only available when the resource is acquired
-        pub const fn pool(&self) -> &Arc<PgPool> {
-            self.get()
-        }
-
-        /// Check pool health
-        ///
-        /// Only available when the resource is acquired
-        pub fn pool_stats(&self) -> PoolStats {
-            let pool = self.get();
-            PoolStats {
-                size: pool.size(),
-                idle: u32::try_from(pool.num_idle()).unwrap_or(u32::MAX),
-                is_closed: pool.is_closed(),
-            }
-        }
-    }
-
-    /// Extension methods for database connection resources
-    impl DatabaseConnection {
-        /// Execute a query using the connection
-        ///
-        /// Only available when the resource is acquired
-        pub async fn execute_query(
-            &mut self,
-            query: &str,
-        ) -> ResourceResult<sqlx::postgres::PgQueryResult> {
-            sqlx::query(query)
-                .execute(&mut **self.get_mut())
-                .await
-                .map_err(|e| ResourceError::InvalidState(format!("Query execution failed: {e}")))
-        }
-
-        // Note: To begin a transaction, use DatabasePool::begin_transaction() instead.
-        // Pool connections should not create their own transactions as this can lead
-        // to lifetime issues and is not the recommended SQLx pattern.
-    }
-
-    /// Extension methods for database transaction resources
-    impl DatabaseTransaction<'_> {
-        /// Execute a query within the transaction
-        ///
-        /// Only available when the resource is acquired
-        pub async fn execute_query(
-            &mut self,
-            query: &str,
-        ) -> ResourceResult<sqlx::postgres::PgQueryResult> {
-            sqlx::query(query)
-                .execute(&mut **self.get_mut())
-                .await
-                .map_err(|e| ResourceError::InvalidState(format!("Transaction query failed: {e}")))
-        }
-
-        /// Commit the transaction
-        ///
-        /// Only available when the resource is acquired
-        /// Consumes the transaction and returns a released resource
-        pub async fn commit(self) -> ResourceResult<Resource<(), states::Released>> {
-            self.into_inner().commit().await.map_err(|e| {
-                ResourceError::ReleaseFailed(format!("Transaction commit failed: {e}"))
-            })?;
-
-            Ok(Resource::new(()))
-        }
-
-        /// Rollback the transaction
-        ///
-        /// Only available when the resource is acquired
-        /// Consumes the transaction and returns a released resource
-        pub async fn rollback(self) -> ResourceResult<Resource<(), states::Released>> {
-            self.into_inner().rollback().await.map_err(|e| {
-                ResourceError::ReleaseFailed(format!("Transaction rollback failed: {e}"))
-            })?;
-
-            Ok(Resource::new(()))
-        }
-    }
-
-    /// Pool statistics for monitoring
-    #[derive(Debug, Clone)]
-    pub struct PoolStats {
-        /// Current pool size
-        pub size: u32,
-        /// Number of idle connections
-        pub idle: u32,
-        /// Whether the pool is closed
-        pub is_closed: bool,
-    }
-
-    /// Factory for creating database resource managers
-    pub struct DatabaseResourceFactory;
-
-    impl DatabaseResourceFactory {
-        /// Create a resource manager from an existing pool
-        pub fn from_pool(pool: Arc<PgPool>) -> DatabaseResourceManager {
-            DatabaseResourceManager::new(pool)
-        }
-
-        /// Create a resource manager with custom health check interval
-        pub fn from_pool_with_health_interval(
-            pool: Arc<PgPool>,
-            health_check_interval: Duration,
-        ) -> DatabaseResourceManager {
-            DatabaseResourceManager::new_with_health_interval(pool, health_check_interval)
-        }
-    }
-}
-
-/// Lock resource implementation with phantom types
-pub mod locking {
-    use super::{states, PhantomData, Resource, ResourceError, ResourceResult};
-    use std::sync::{Arc, Mutex, MutexGuard};
-
-    /// A mutex lock wrapped in resource management
-    pub type MutexResource<T> = Resource<Arc<Mutex<T>>, states::Acquired>;
-
-    /// A mutex guard wrapped in resource management
-    ///
-    /// This ensures the guard is only used while the lock is held
-    pub struct MutexGuardResource<'a, T> {
-        guard: MutexGuard<'a, T>,
-        _phantom: PhantomData<states::Acquired>,
-    }
-
-    impl<'a, T> MutexGuardResource<'a, T> {
-        /// Create a new mutex guard resource
-        const fn new(guard: MutexGuard<'a, T>) -> Self {
-            Self {
-                guard,
-                _phantom: PhantomData,
-            }
-        }
-
-        /// Access the protected data
-        pub fn get(&self) -> &T {
-            &self.guard
-        }
-
-        /// Access the protected data mutably
-        pub fn get_mut(&mut self) -> &mut T {
-            &mut self.guard
-        }
-    }
-
-    /// Extension methods for mutex resources
-    impl<T> MutexResource<T> {
-        /// Acquire the mutex lock
-        ///
-        /// Returns a guard resource that enforces lock is held
-        pub fn lock(&self) -> ResourceResult<MutexGuardResource<'_, T>> {
-            let guard = self
-                .get()
-                .lock()
-                .map_err(|e| ResourceError::Poisoned(format!("Mutex poisoned: {e}")))?;
-            Ok(MutexGuardResource::new(guard))
-        }
-
-        /// Try to acquire the mutex lock without blocking
-        ///
-        /// Returns None if the lock is currently held
-        pub fn try_lock(&self) -> ResourceResult<Option<MutexGuardResource<'_, T>>> {
-            match self.get().try_lock() {
-                Ok(guard) => Ok(Some(MutexGuardResource::new(guard))),
-                Err(std::sync::TryLockError::WouldBlock) => Ok(None),
-                Err(std::sync::TryLockError::Poisoned(e)) => {
-                    Err(ResourceError::Poisoned(format!("Mutex poisoned: {e}")))
-                }
-            }
-        }
-    }
-
-    /// Create a mutex resource
-    pub fn create_mutex_resource<T>(data: T) -> MutexResource<T> {
-        Resource::new(Arc::new(Mutex::new(data)))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::states::*;
-    use super::{
-        global_leak_detector, locking, IsAcquired, IsReleasable, ManagedResource, Resource,
-        ResourceError, ResourceExt, ResourceLeakDetector, ResourceScope, TimedResourceGuard,
-    };
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::Arc;
-    use std::time::Duration;
-    use tokio::time::{sleep, timeout};
-
-    #[test]
-    fn test_resource_state_transitions() {
-        // Start with initializing resource
-        let initializing = Resource::<String, Initializing>::new("test".to_string());
-
-        // Can transition to acquired
-        let acquired = initializing.mark_acquired();
-        assert_eq!(acquired.get(), "test");
-
-        // Can transition to failed
-        let failed = acquired.mark_failed();
-
-        // Failed can be recovered
-        let recovered = failed.recover().unwrap();
-        assert_eq!(recovered.get(), "test");
-
-        // Can be released
-        let released = recovered.release().unwrap();
-        assert!(released.is_released());
-    }
-
-    #[test]
-    fn test_resource_scope() {
-        let resource = Resource::<String, Acquired>::new("test".to_string());
-        let mut scope = ResourceScope::new(resource);
-
-        // Can access resource in scope
-        scope.with_resource(|r| {
-            assert_eq!(r.get(), "test");
-        });
-
-        // Check scope state
-        assert!(!scope.is_released());
-        assert!(!scope.is_leaked());
-
-        // Manual release works
-        scope.release().unwrap();
-    }
-
-    #[test]
-    fn test_resource_scope_automatic_cleanup() {
-        let leaked_flag = Arc::new(AtomicUsize::new(0));
-
-        // Create scope and let it drop without explicit release
-        {
-            let resource = Resource::<String, Acquired>::new("test".to_string());
-            let _scope = ResourceScope::new(resource);
-
-            // Increment counter when scope is created
-            leaked_flag.store(1, Ordering::SeqCst);
-
-            // Scope will be dropped here without explicit release
-        }
-
-        // Verify scope was created (this tests that the test setup works)
-        assert_eq!(leaked_flag.load(Ordering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn test_timed_resource_guard() {
-        let resource = Resource::<String, Acquired>::new("test".to_string());
-        let timeout_duration = Duration::from_millis(100);
-        let guard = TimedResourceGuard::new(resource, timeout_duration);
-
-        // Initially should be available
-        assert!(guard.get().is_some());
-        assert!(!guard.is_timed_out());
-        assert!(guard.time_remaining().is_some());
-
-        // Wait for timeout
-        sleep(Duration::from_millis(150)).await;
-
-        // Should be timed out
-        assert!(guard.is_timed_out());
-        assert!(guard.get().is_none());
-        assert!(guard.time_remaining().is_none());
-
-        // Release should fail with timeout error
-        match guard.release() {
-            Err(ResourceError::Timeout { .. }) => {
-                // Expected timeout error
-            }
-            other => panic!("Expected timeout error, got: {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_timed_resource_guard_early_release() {
-        let resource = Resource::<String, Acquired>::new("test".to_string());
-        let guard = TimedResourceGuard::new(resource, Duration::from_secs(10));
-
-        // Should be available
-        assert!(guard.get().is_some());
-        assert!(!guard.is_timed_out());
-
-        // Release before timeout
-        guard.release().unwrap();
-    }
-
-    #[test]
-    fn test_mutex_resource() {
-        let mutex_resource = locking::create_mutex_resource(42i32);
-
-        // Can acquire lock
-        let guard = mutex_resource.lock().unwrap();
-        assert_eq!(*guard.get(), 42);
-
-        // Lock is exclusive
-        assert!(mutex_resource.try_lock().unwrap().is_none()); // Should fail to acquire
-
-        // Release first lock
-        drop(guard);
-
-        // Now should be able to acquire
-        assert!(mutex_resource.try_lock().unwrap().is_some());
-    }
-
-    #[test]
-    fn test_mutex_resource_mutable_access() {
-        let mutex_resource = locking::create_mutex_resource(42i32);
-
-        // Acquire lock and modify data
-        {
-            let mut guard = mutex_resource.lock().unwrap();
-            *guard.get_mut() = 100;
-        }
-
-        // Verify modification
-        assert_eq!(*mutex_resource.lock().unwrap().get(), 100);
-    }
-
-    #[test]
-    fn test_resource_leak_detector() {
-        let detector = ResourceLeakDetector::new();
-
-        // Initially empty
-        let initial_stats = detector.get_stats();
-        assert_eq!(initial_stats.total_active, 0);
-        assert!(initial_stats.by_type.is_empty());
-
-        // Register some resources
-        detector.register_acquisition("res1", "DatabasePool", Some("test_location".to_string()));
-        detector.register_acquisition("res2", "DatabasePool", None);
-        detector.register_acquisition("res3", "MutexLock", None);
-
-        let stats = detector.get_stats();
-        assert_eq!(stats.total_active, 3);
-        assert_eq!(stats.by_type.get("DatabasePool"), Some(&2));
-        assert_eq!(stats.by_type.get("MutexLock"), Some(&1));
-
-        // Release one resource
-        detector.register_release("res1");
-
-        let stats = detector.get_stats();
-        assert_eq!(stats.total_active, 2);
-        assert_eq!(stats.by_type.get("DatabasePool"), Some(&1));
-
-        // Find potential leaks (none yet since resources are new)
-        let leaks = detector.find_potential_leaks(Duration::from_secs(1));
-        assert!(leaks.is_empty());
-
-        // Clean up
-        detector.register_release("res2");
-        detector.register_release("res3");
-
-        let final_stats = detector.get_stats();
-        assert_eq!(final_stats.total_active, 0);
-    }
-
-    #[test]
-    fn test_global_leak_detector() {
-        let initial_count = global_leak_detector().get_stats().total_active;
-
-        // Register a resource
-        global_leak_detector().register_acquisition("global_test", "TestResource", None);
-
-        let stats = global_leak_detector().get_stats();
-        assert_eq!(stats.total_active, initial_count + 1);
-
-        // Release the resource
-        global_leak_detector().register_release("global_test");
-
-        let final_stats = global_leak_detector().get_stats();
-        assert_eq!(final_stats.total_active, initial_count);
-    }
-
-    #[test]
-    fn test_managed_resource() {
-        let initial_count = global_leak_detector().get_stats().total_active;
-
-        // Create managed resource
-        let resource = Resource::<String, Acquired>::new("test".to_string());
-        let managed = ManagedResource::new(resource, "TestResource");
-
-        // Should be tracked
-        let stats = global_leak_detector().get_stats();
-        assert!(stats.total_active > initial_count);
-        assert!(stats.by_type.contains_key("TestResource"));
-
-        // Can access inner resource
-        assert!(managed.get().is_some());
-
-        // Taking resource should work and update tracking
-        let taken = managed.take();
-        assert!(taken.is_some());
-
-        // Should be untracked now
-        let final_stats = global_leak_detector().get_stats();
-        assert_eq!(final_stats.total_active, initial_count);
-    }
-
-    #[test]
-    fn test_managed_resource_drop_cleanup() {
-        let initial_count = global_leak_detector().get_stats().total_active;
-
-        // Create managed resource and drop it
-        {
-            let resource = Resource::<String, Acquired>::new("test".to_string());
-            let _managed = ManagedResource::new(resource, "TestResource");
-
-            // Should be tracked
-            let stats = global_leak_detector().get_stats();
-            assert!(stats.total_active > initial_count);
-        } // Drop happens here
-
-        // Should be untracked after drop
-        let final_stats = global_leak_detector().get_stats();
-        assert_eq!(final_stats.total_active, initial_count);
-    }
-
-    #[test]
-    fn test_resource_extension_traits() {
-        let resource = Resource::<String, Acquired>::new("test".to_string());
-
-        // Test managed extension
-        let managed = resource.managed("TestResource");
-        assert!(managed.get().is_some());
-
-        let resource2 = Resource::<String, Acquired>::new("test2".to_string());
-
-        // Test scoped extension
-        let scope = resource2.scoped();
-        assert!(!scope.is_released());
-
-        let resource3 = Resource::<String, Acquired>::new("test3".to_string());
-
-        // Test timed extension
-        let timed = resource3.with_timeout(Duration::from_secs(1));
-        assert!(timed.get().is_some());
-    }
-
-    #[tokio::test]
-    async fn test_resource_state_machine_invalid_transitions() {
-        // Test that certain state transitions are not allowed at compile time
-
-        let released = Resource::<String, Released>::new("test".to_string());
-        assert!(released.is_released());
-
-        // These operations should not compile (verified by compiler):
-        // released.get(); // ❌ Cannot access released resource
-        // released.release(); // ❌ Cannot release already released resource
-        // released.mark_failed(); // ❌ Cannot fail released resource
-    }
-
-    #[tokio::test]
-    async fn test_concurrent_resource_access() {
-        let resource = locking::create_mutex_resource(0i32);
-        let resource = Arc::new(resource);
-
-        // Spawn multiple tasks that increment the counter
-        let mut handles = vec![];
-        for _ in 0..10 {
-            let resource_clone = resource.clone();
-            let handle = tokio::spawn(async move {
-                let mut guard = resource_clone.lock().unwrap();
-                let current = *guard.get();
-                *guard.get_mut() = current + 1;
-                // Lock is released when guard is dropped
-            });
-            handles.push(handle);
-        }
-
-        // Wait for all tasks to complete
-        for handle in handles {
-            handle.await.unwrap();
-        }
-
-        // Verify final value
-        assert_eq!(*resource.lock().unwrap().get(), 10);
-    }
-
-    #[tokio::test]
-    async fn test_resource_timeout_behavior() {
-        let resource = Resource::<String, Acquired>::new("test".to_string());
-        let guard = TimedResourceGuard::new(resource, Duration::from_millis(50));
-
-        // Test that timeout actually works
-        let result = timeout(Duration::from_millis(100), async {
-            while !guard.is_timed_out() {
-                sleep(Duration::from_millis(10)).await;
-            }
-        })
-        .await;
-
-        assert!(result.is_ok(), "Timeout should have occurred within 100ms");
-        assert!(guard.is_timed_out());
-    }
-
-    #[test]
-    fn test_resource_error_types() {
-        // Test different error variants
-        let acquisition_error = ResourceError::AcquisitionFailed("test".to_string());
-        assert!(matches!(
-            acquisition_error,
-            ResourceError::AcquisitionFailed(_)
-        ));
-
-        let release_error = ResourceError::ReleaseFailed("test".to_string());
-        assert!(matches!(release_error, ResourceError::ReleaseFailed(_)));
-
-        let invalid_state_error = ResourceError::InvalidState("test".to_string());
-        assert!(matches!(
-            invalid_state_error,
-            ResourceError::InvalidState(_)
-        ));
-
-        let timeout_error = ResourceError::Timeout {
-            duration: Duration::from_secs(1),
-        };
-        assert!(matches!(timeout_error, ResourceError::Timeout { .. }));
-
-        let poisoned_error = ResourceError::Poisoned("test".to_string());
-        assert!(matches!(poisoned_error, ResourceError::Poisoned(_)));
-    }
-
-    #[test]
-    fn test_sealed_trait_pattern() {
-        // Test that our sealed traits work correctly
-        // This is mostly a compile-time test
-
-        fn test_acquired<S: IsAcquired>(_state: std::marker::PhantomData<S>) {
-            // This function should only accept acquired states
-        }
-
-        fn test_releasable<S: IsReleasable>(_state: std::marker::PhantomData<S>) {
-            // This function should only accept releasable states
-        }
-
-        // These should compile
-        test_acquired(std::marker::PhantomData::<Acquired>);
-        test_releasable(std::marker::PhantomData::<Acquired>);
-        test_releasable(std::marker::PhantomData::<Failed>);
-
-        // These should NOT compile (verified by external compile tests):
-        // test_acquired(std::marker::PhantomData::<Released>);
-        // test_releasable(std::marker::PhantomData::<Released>);
-    }
-
-    #[test]
-    fn test_compilation_errors() {
-        // These tests verify that certain operations don't compile
-        // They are written as compile_fail tests to document the expected behavior
-
-        // Cannot use released resource
-        let _resource = Resource::<String, Released>::new("test".to_string());
-        // resource.get(); // ❌ This should not compile
-
-        // Cannot release an already released resource
-        // resource.release(); // ❌ This should not compile
-
-        // Cannot mark released resource as failed
-        // resource.mark_failed(); // ❌ This should not compile
     }
 }
